Create SaritaPhadiaCleanup_08282015.R
Sarita: Took Cecilia's code and compiled the existing Phadia data from our machines along with the data in Redcap.  I made the code available in case anyone wants to get an idea of some of the issues.  Warning: not beautiful code.

Identified a few issues:
1. Handling repeated runs on the same sample, usually with discordant values: 
ie: 25BL: there was a previous aIgE and now another one that we did, they are off by 100

2. Misnamed visits, as part of the sampleID.  It required some additional code as otherwise it would be difficult to merge on…probably not a problem for anyone else.

3. I think the database still needs some cleanup---I'm still getting duplicate errors.  

4. At the end, it didn't make sense for me to sink enough time to write a loop which would meaningfully merge the phadia from our machine and phadia from redcap data, so I ended up having to handcurate it.  By the time we are finished, and do get actual values for the measurements of IgE which are over the maximum limit, this won't be a problem anymore.  I will also make my more merged file with what I think is a reasonable start available.